# Eager load existing notifications when syncing