A quick attempt to speed up the syncing of notifications, especially for new users, this eager loads the existing notifications for a user rather than letting `find_or_initialize_by` do it for every item in the API response.

Quick testing in development shows around ~20 extra sql queries killed, saving about 200ms on my account:

    Before: Completed 302 Found in 1541ms (ActiveRecord: 64.4ms | SQL count: 105)
    After:  Completed 302 Found in 1331ms (ActiveRecord: 42.4ms | SQL count: 82)

There's also some other savings to be made by inserting/upserting all the records in one go as well in future.